requirements.txt
flask==2.3.3
werkzeug==2.3.7
jinja2==3.1.2
markupsafe==2.1.3
itsdangerous==2.1.2
python-dotenv==1.0.0
requests==2.31.0
pyjwt==2.8.0
bcrypt==4.0.1
bleach==6.0.0
pillow==10.0.0
pytest==7.4.0
flask-socketio==5.3.4
gevent==23.7.0
gevent-websocket==0.10.1
python-dateutil==2.8.2 